+++ src/main/kotlin/com/personio/synthetics/builder/SyntheticTestBuilder.kt
@@ -34,7 +34,6 @@ abstract class SyntheticTestBuilder(
 ) {
     protected var parameters: SyntheticTestParameters
     protected var options: SyntheticsTestOptions
-    protected var locations: List<String> = defaults.runLocations
     protected var status: SyntheticsTestPauseStatus = SyntheticsTestPauseStatus.PAUSED
 
     init {
@@ -173,8 +172,8 @@ abstract class SyntheticTestBuilder(
      * Allowed minimum location failed is between 1 and the number of locations where the test is configured to run
      */
     fun minLocationFailed(minLocationFailed: Long) {
-        require(minLocationFailed in 1..locations.count()) {
-            "Minimum location failed should be between 1 and the number of locations where the test is configured to run: ${locations.count()}."
+        require(minLocationFailed in 1..parameters.locations.count()) {
+            "Minimum location failed should be between 1 and the number of locations where the test is configured to run: ${parameters.locations.count()}."
         }
         options.minLocationFailed = minLocationFailed
     }
@@ -192,8 +191,8 @@ abstract class SyntheticTestBuilder(
      * @param monitorPriority The monitor priority of the test
      * Allowed monitor priority is one of [1, 2, 3, 4, 5]
      */
-    fun monitorPriority(monitorPriorities: MonitorPriority) {
-        options.monitorPriority = monitorPriorities.priorityValue
+    fun monitorPriority(monitorPriority: MonitorPriority) {
+        options.monitorPriority = monitorPriority.priorityValue
     }
 
     /**